Scene changes, Added characters, Audience headbop prototype
Changed the main scene by adding a lot of textures, assets and characters.
A lot of character assets and scene assets has been added to the project.
Made headbop mechanic of the audience (Only present on one NPC so far, but it's just base slave set-up work now to add others)