AudioManager + Scene Edit + AudienceAnimations
Added animations to some audience. Still have problems with this since the imported animations are generic and not humanoid, I cannot use avatar-masking on the head of the audience to get that Headbop effect we have talked about.

Added more assets and things to the scene.

Added an AudioManager which can be used for sounds that are not dependent on 3D audio localization.